refactor(router): split routes/inference.rs (q1 wave 5/5 — final)
Extract inference + foundry handlers from routes/mod.rs into routes/inference.rs
(~1850 LOC). Pure mechanical split — zero drift on this wave.

## What moved

- Route builders: inference_routes, foundry_agent_routes, foundry_standalone_routes
- 6 inference handlers: chat_completions, completions, responses, embeddings,
  images_generations, images_generations_v1
- 3 foundry handlers: list_models, list_deployments, foundry_proxy
- Helpers: uuid_v4, chat_to_responses_body, responses_to_chat_body
- The #[cfg(test)] mod tests block (5 body-transform tests)

## What stayed in mod.rs (336 LOC, down from 4890)

- AppState struct + upstream_config() method
- DEPRECATED_ADMIN_HEADER_WARNED static + extract_admin_token()
- health_routes/metrics_routes/admin_routes + healthz/readyz/metrics/
  admin_get_model/admin_set_model handlers
- mod <domain>; pub use ... re-exports for the 5 submodules

## Byte-equivalence proof (cumulative q1 waves 1-5)

Drift checker reports exactly ONE intentional mutation across the entire
q1 refactor — in handoff_protected_routes (wave 4), documented in
commit a277a14 as a namespace-resolution fix.

All other 84 fn+data items (inference_routes, foundry_*_routes,
chat_completions, completions, responses, embeddings, images_generations,
images_generations_v1, list_models, list_deployments, foundry_proxy,
uuid_v4, chat_to_responses_body, responses_to_chat_body, all 5 test
bodies, everything in egress/governance/mesh/handoff) are byte-identical
to main.

## Final layout

    inference-router/src/routes/
    ├── mod.rs            336 LOC  (AppState, admin, health, metrics)
    ├── inference.rs     1865 LOC  (inference + foundry proxy)
    ├── handoff.rs       1597 LOC  (spawn + handoff)
    ├── governance.rs     418 LOC  (AGT trust/policy/audit)
    ├── mesh.rs           421 LOC  (AgentMesh relay + registry)
    └── egress.rs         347 LOC  (proxy + DNS)

Total 4984 LOC across 6 files; original routes.rs was 4890 LOC in 1 file.
Net +94 LOC = mod-declarations + imports-per-submodule boilerplate.

## Verification

- cargo test --all: 275/275 passed
- cd cli && npx vitest run: 240 passed + 2 skipped (242 total)
- cd mesh-plugin && npm test: 36/36 passed
- cargo clippy --all-targets -- -D warnings: clean
- Drift: 1 intentional mutation, documented in wave 4 commit
